fix(matching-contexts): Fix all matching contexts being fetched after a subscribe/unsubscribe action.
The bug was probably introduced while working on the profiles app.
The matching contexts are not used anymore on the profile app, so this part was just removed.
