+++ src/app/actions/refreshMatchingContexts.ts
@@ -1,18 +1,13 @@
 import { MatchingContext } from 'app/lmem/matchingContext';
 import { BaseAction, ErrorAction } from '.';
 import { Level } from '../utils/Logger';
-import { ContributorId } from 'app/lmem/contributor';
 
 export const REFRESH_MATCHING_CONTEXTS = 'REFRESH_MATCHING_CONTEXTS';
 export interface RefreshMatchingContextsAction extends BaseAction {
   type: typeof REFRESH_MATCHING_CONTEXTS;
-  payload: { contributors: ContributorId[] };
 }
-export const refreshMatchingContexts = (
-  contributors: ContributorId[] = []
-): RefreshMatchingContextsAction => ({
-  type: REFRESH_MATCHING_CONTEXTS,
-  payload: { contributors }
+export const refreshMatchingContexts = (): RefreshMatchingContextsAction => ({
+  type: REFRESH_MATCHING_CONTEXTS
 });
 
 export const UPDATE_MATCHING_CONTEXTS = 'api/UPDATE_MATCHING_CONTEXTS';
